+++ .storybook/components/StorybookColorSwatches/StorybookColorSwatches.tsx
@@ -5,14 +5,14 @@ import { Flex } from "@jobber/components/Flex";
 import { Text } from "@jobber/components/Text";
 import { Content } from "@jobber/components/Content";
 import { showToast } from "@jobber/components/Toast";
-import styles from "./ColorSwatches.module.css";
+import styles from "./StorybookColorSwatches.module.css";
 
-interface ColorSwatchesProps {
+interface StorybookColorSwatchesProps {
   readonly colors: string[];
   readonly description?: string;
 }
 
-export function ColorSwatches({ colors }: ColorSwatchesProps) {
+export function StorybookColorSwatches({ colors }: StorybookColorSwatchesProps) {
   return (
     <Content>
       {colors.map((color: string) => (