+++ packages/components/src/Combobox/Combobox.types.ts
@@ -60,8 +60,43 @@ export interface ComboboxProps {
   readonly loading?: boolean;
 }
 
+export interface ComboboxCustomActivatorProps {
+  /**
+   * Required to describe the expanded state of the Combobox.
+   * Automatically updates as open state is toggled.
+   */
+  ariaExpanded: boolean;
+
+  /**
+   * Required to describe the relationship between the toggle-able Combobox Menu and the activator.
+   */
+  ariaControls: string;
+
+  /**
+   * The aria-label attribute for the Combobox activator.
+   */
+  ariaLabel?: string;
+
+  /**
+   * Required to describe the interactive element which toggles the Combobox's visibility.
+   */
+  role: "combobox";
+
+  /**
+   * Method to open the Combobox. Closing is handled by the Combobox itself.
+   */
+  open: () => void;
+}
+
+export type ComboboxActivatorAccessibility = Omit<
+  ComboboxCustomActivatorProps,
+  "open"
+>;
+
 export interface ComboboxActivatorProps {
-  readonly children: React.ReactElement;
+  readonly children:
+    | React.ReactElement
+    | ((props: ComboboxCustomActivatorProps) => React.ReactElement);
 }
 
 export interface ComboboxTriggerProps
@@ -147,11 +182,6 @@ export interface ComboboxContentProps {
    * Is the Combobox open
    */
   readonly open: boolean;
-
-  /**
-   * Setter for the open state of the Combobox.
-   */
-  readonly setOpen: (open: boolean) => void;
 
   /**
    * The full set of options for the Combobox in the shape of data, not elements.
@@ -278,6 +308,12 @@ export interface ComboboxActionProps {
   readonly visible?:
     | boolean
     | ((options: ComboboxActionCallbackOptions) => boolean);
+
+  /**
+   * Keeps the Combobox open when the action is clicked.
+   * @default false
+   */
+  readonly keepOpenOnClick?: boolean;
 }
 
 export interface ComboboxActionCallbackOptions {