+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContentSearch/ComboboxContentSearch.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef } from "react";
 import { Icon } from "@jobber/components/Icon";
-import styles from "./ComboboxContentSearch.css";
+import styles from "./ComboboxContentSearch.module.css";
 import { ComboboxSearchProps } from "../../../Combobox.types";
 
 export function ComboboxContentSearch(props: ComboboxSearchProps): JSX.Element {