+++ packages/components/src/DatePicker/DatePickerCustomHeader.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ReactDatePickerCustomHeaderProps } from "react-datepicker";
-import styles from "./DatePicker.css";
+import styles from "./DatePicker.module.css";
 import { Typography } from "../Typography";
 import { Button } from "../Button";