+++ packages/components/src/Grid/GridCell/GridCell.test.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import { GRID_CELL_TEST_ID, InternalGridCell } from ".";
+import { GRID_CELL_TEST_ID, GridCell } from ".";
 
 describe("GridCell", () => {
   it("should add a css variable inline styling", () => {
     render(
-      <InternalGridCell size={{ xs: 12, sm: 10, md: 8, lg: 6, xl: 4 }}>
+      <GridCell size={{ xs: 12, sm: 10, md: 8, lg: 6, xl: 4 }}>
         Even numbers baby!
-      </InternalGridCell>,
+      </GridCell>,
     );
 
     const element = screen.getByTestId(GRID_CELL_TEST_ID);
@@ -22,20 +22,14 @@ describe("GridCell", () => {
 
   describe("Inferred sizing", () => {
     it("should set xs as 12 if it's not declared", () => {
-      render(
-        <InternalGridCell size={{ sm: 2 }}>No XS for you!</InternalGridCell>,
-      );
+      render(<GridCell size={{ sm: 2 }}>No XS for you!</GridCell>);
 
       expect(getCssVar("--gridCell--size-xs")).toBe("12");
       expect(getCssVar("--gridCell--size-sm")).toBe("2");
     });
 
     it("should use the previous value for sizing", () => {
-      render(
-        <InternalGridCell size={{ xs: 6, md: 7 }}>
-          Size me from prev
-        </InternalGridCell>,
-      );
+      render(<GridCell size={{ xs: 6, md: 7 }}>Size me from prev</GridCell>);
 
       expect(getCssVar("--gridCell--size-xs")).toBe("6");
       expect(getCssVar("--gridCell--size-sm")).toBe("6");
@@ -45,11 +39,7 @@ describe("GridCell", () => {
     });
 
     it("should use xs sizing for all sizing when only xs is declared", () => {
-      render(
-        <InternalGridCell size={{ xs: 10 }}>
-          You all get a 10!
-        </InternalGridCell>,
-      );
+      render(<GridCell size={{ xs: 10 }}>You all get a 10!</GridCell>);
 
       expect(getCssVar("--gridCell--size-xs")).toBe("10");
       expect(getCssVar("--gridCell--size-sm")).toBe("10");